chore(web): mock pool attempts in records stories
