+++ web/src/components/InvocationRecordsTable.stories.tsx
@@ -1,7 +1,8 @@
+import { useEffect, useMemo, useRef, type ReactNode } from 'react'
 import type { Meta, StoryObj } from '@storybook/react-vite'
 import { I18nProvider } from '../i18n'
 import { InvocationRecordsTable } from './InvocationRecordsTable'
-import { STORYBOOK_INVOCATION_RECORDS } from './invocationRecordsStoryFixtures'
+import { createStoryPoolAttemptsByInvokeId, STORYBOOK_INVOCATION_RECORDS } from './invocationRecordsStoryFixtures'
 
 function StorySurface({ children }: { children: React.ReactNode }) {
   return (
@@ -9,6 +10,50 @@ function StorySurface({ children }: { children: React.ReactNode }) {
       <div className="mx-auto w-full max-w-[1440px]">{children}</div>
     </div>
   )
+}
+
+function jsonResponse(payload: unknown) {
+  return new Response(JSON.stringify(payload), {
+    status: 200,
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  })
+}
+
+function StorybookPoolAttemptsMock({ children, records }: { children: ReactNode; records: typeof STORYBOOK_INVOCATION_RECORDS }) {
+  const originalFetchRef = useRef<typeof window.fetch | null>(null)
+  const poolAttemptsByInvokeId = useMemo(() => createStoryPoolAttemptsByInvokeId(records), [records])
+
+  if (typeof window !== 'undefined' && !originalFetchRef.current) {
+    originalFetchRef.current = window.fetch.bind(window)
+
+    const mockedFetch: typeof window.fetch = async (input, init) => {
+      const requestUrl = typeof input === 'string' ? input : input instanceof URL ? input.toString() : input.url
+      const url = new URL(requestUrl, window.location.origin)
+      const poolAttemptsMatch = url.pathname.match(/^\/api\/invocations\/([^/]+)\/pool-attempts$/)
+
+      if (poolAttemptsMatch) {
+        const invokeId = decodeURIComponent(poolAttemptsMatch[1] ?? '')
+        return jsonResponse(poolAttemptsByInvokeId[invokeId] ?? [])
+      }
+
+      return (originalFetchRef.current as typeof window.fetch)(input, init)
+    }
+
+    window.fetch = mockedFetch
+  }
+
+  useEffect(() => {
+    return () => {
+      if (typeof window !== 'undefined' && originalFetchRef.current) {
+        window.fetch = originalFetchRef.current
+        originalFetchRef.current = null
+      }
+    }
+  }, [])
+
+  return <>{children}</>
 }
 
 const meta = {
@@ -20,9 +65,11 @@ const meta = {
   decorators: [
     (Story) => (
       <I18nProvider>
-        <StorySurface>
-          <Story />
-        </StorySurface>
+        <StorybookPoolAttemptsMock records={STORYBOOK_INVOCATION_RECORDS}>
+          <StorySurface>
+            <Story />
+          </StorySurface>
+        </StorybookPoolAttemptsMock>
       </I18nProvider>
     ),
   ],
