+++ web/src/components/RecordsPage.stories.tsx
@@ -4,13 +4,13 @@ import { expect, userEvent, waitFor, within } from 'storybook/test'
 import { I18nProvider } from '../i18n'
 import type {
   ApiInvocation,
-  ApiPoolUpstreamRequestAttempt,
   InvocationRecordsQuery,
   InvocationSortBy,
   InvocationSortOrder,
 } from '../lib/api'
 import RecordsPage from '../pages/Records'
 import {
+  createStoryPoolAttemptsByInvokeId,
   createStoryInvocationRecordsResponse,
   createStoryInvocationRecordsSummary,
   STORYBOOK_INVOCATION_RECORDS,
@@ -185,88 +185,6 @@ function buildSuggestions(records: ApiInvocation[]) {
   }
 }
 
-function resolveFinalAttemptStatus(record: ApiInvocation): ApiPoolUpstreamRequestAttempt['status'] {
-  if (record.status === 'success') return 'success'
-  if (record.poolAttemptTerminalReason === 'budget_exhausted_final') return 'budget_exhausted_final'
-  if (record.failureClass === 'client_failure') return 'http_failure'
-  return 'transport_failure'
-}
-
-function createPoolAttemptFixtures(records: ApiInvocation[]) {
-  const attemptsByInvokeId: Record<string, ApiPoolUpstreamRequestAttempt[]> = {}
-
-  records.forEach((record, recordIndex) => {
-    if (record.routeMode !== 'pool') return
-
-    const attemptTotal = Math.max(1, record.poolAttemptCount ?? 1)
-    const distinctTotal = Math.max(1, Math.min(attemptTotal, record.poolDistinctAccountCount ?? attemptTotal))
-    const finalAccountId =
-      typeof record.upstreamAccountId === 'number' && Number.isFinite(record.upstreamAccountId)
-        ? Math.trunc(record.upstreamAccountId)
-        : 100 + recordIndex * 10 + distinctTotal
-    const baseAccountId = finalAccountId - (distinctTotal - 1)
-    const finalStatus = resolveFinalAttemptStatus(record)
-
-    attemptsByInvokeId[record.invokeId] = Array.from({ length: attemptTotal }, (_, attemptIndex) => {
-      const isLastAttempt = attemptIndex === attemptTotal - 1
-      const distinctAccountIndex = Math.min(distinctTotal, attemptIndex + 1)
-      const sameAccountRetryIndex = attemptIndex < distinctTotal ? 1 : attemptIndex - distinctTotal + 2
-      const upstreamAccountId = isLastAttempt ? finalAccountId : baseAccountId + distinctAccountIndex - 1
-      const upstreamAccountName =
-        isLastAttempt && record.upstreamAccountName?.trim()
-          ? record.upstreamAccountName
-          : `Pool Candidate ${upstreamAccountId}`
-      const status = isLastAttempt ? finalStatus : attemptIndex % 2 === 0 ? 'transport_failure' : 'http_failure'
-      const startedAtMs = Date.parse(record.occurredAt) - (attemptTotal - attemptIndex) * 900
-      const finishedAtMs = startedAtMs + 240 + attemptIndex * 90
-
-      return {
-        id: record.id * 100 + attemptIndex + 1,
-        invokeId: record.invokeId,
-        occurredAt: record.occurredAt,
-        endpoint: record.endpoint ?? '/v1/responses',
-        attemptIndex: attemptIndex + 1,
-        distinctAccountIndex,
-        sameAccountRetryIndex,
-        status,
-        httpStatus:
-          status === 'success' ? 200 : status === 'http_failure' ? (isLastAttempt ? 400 : 429) : null,
-        failureKind:
-          status === 'success'
-            ? null
-            : status === 'http_failure'
-              ? isLastAttempt
-                ? record.failureKind ?? 'invalid_request'
-                : 'rate_limit'
-              : 'connect_timeout',
-        errorMessage:
-          status === 'success'
-            ? null
-            : status === 'http_failure'
-              ? isLastAttempt
-                ? record.errorMessage ?? 'upstream rejected request'
-                : 'upstream returned 429'
-              : 'forward proxy connect timeout',
-        connectLatencyMs: status === 'transport_failure' ? 160 + attemptIndex * 20 : 55 + attemptIndex * 12,
-        firstByteLatencyMs:
-          status === 'success' ? (record.tUpstreamTtfbMs ?? 180) : status === 'http_failure' ? 210 + attemptIndex * 18 : null,
-        streamLatencyMs:
-          status === 'success' && typeof record.tTotalMs === 'number' && Number.isFinite(record.tTotalMs)
-            ? Math.max(80, record.tTotalMs - (record.tUpstreamTtfbMs ?? 180))
-            : null,
-        upstreamRequestId: `${record.invokeId}-attempt-${attemptIndex + 1}`,
-        startedAt: new Date(startedAtMs).toISOString(),
-        finishedAt: new Date(finishedAtMs).toISOString(),
-        createdAt: new Date(finishedAtMs).toISOString(),
-        upstreamAccountId,
-        upstreamAccountName,
-      }
-    })
-  })
-
-  return attemptsByInvokeId
-}
-
 function jsonResponse(payload: unknown) {
   return new Response(JSON.stringify(payload), {
     status: 200,
@@ -292,7 +210,7 @@ function StorybookRecordsPageMock({
   const originalFetchRef = useRef<typeof window.fetch | null>(null)
   const originalSetIntervalRef = useRef<typeof window.setInterval | null>(null)
   const invocationSearchCountRef = useRef(0)
-  const poolAttemptsByInvokeId = useMemo(() => createPoolAttemptFixtures(records), [records])
+  const poolAttemptsByInvokeId = useMemo(() => createStoryPoolAttemptsByInvokeId(records), [records])
 
   const maybeDelayRefresh = async () => {
     if (refreshDelayMs <= 0) return
